remove children device structures from sensehat struct
each child device now allocates their own structure
