+++ sensehat.h
@@ -12,7 +12,6 @@
 
 #ifndef __LINUX_MFD_SENSEHAT_H_
 #define __LINUX_MFD_SENSEHAT_H_
-#include <linux/miscdevice.h>
 
 #define SENSEDISP_IOC_MAGIC 0xF1
 
@@ -24,21 +23,6 @@ struct sensehat {
 	struct device *dev;
 	struct i2c_client *i2c_client;
 	struct regmap *regmap;
-
-	/* Client devices */
-	struct sensehat_joystick {
-		struct platform_device *pdev;
-		struct input_dev *keys_dev;
-		unsigned long prev_states;
-	} joystick;
-
-	struct sensehat_display {
-		struct platform_device *pdev;
-		struct miscdevice mdev;
-		struct mutex rw_mtx;
-		u8 gamma[32];
-		u8 vmem[192];
-	} display;
 };
 
 enum gamma_preset {
